requirements.txt
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
kafka-python==2.0.2
PyAudio==0.2.13
pycodestyle==2.10.0
requests==2.28.2
SpeechRecognition==3.9.0
tomli==2.0.1
urllib3==1.26.14
